services/auth/requirements.txt
flask==3.1.0
psycopg2-binary==2.9.10
bcrypt==4.2.1
PyJWT==2.10.1
python-dotenv==1.0.1
